feat: work with most files created by java that have members 4294967295 bytes long
We can't up front be sure of the format of the data descriptor in all cases. So
we inch our way through the stream checking what we think is the data
descriptor against the known compressed and uncompressed size of the data
itself. If there is a match: that's "probably" the data descriptor.

The issue is that we could have cases where the we get a match just because the
values happen to match rather than it being a correct match. Or even, because
of errors in the stream, not matching when it should have done. The cases are
not likely, but so far I don't see a reason why they're impossible.