feat: improve detection of data descriptor
The data descriptor is a section after the end of the the member data in some
ZIP files. Whether or not there is a data descriptor is always known up front,
but the exact format is not. So we have to use a heuristic to detect it. There
is unfortunately a theorical possibility that as we walk the stream we would
match too early because values just happen to match.

It's already fairly unlikely, but this commit reduces the chance even more by
ensuring that after what we think is the data descriptor is what seems to be
another member file, or the central directory header.

Because this requires a "peek" ahead in the stream, we need a way of putting
data back to be read again by later code.