feat: allow disabling of zip64
This is to allow stream-unzip to be used in the testing of libraries that
should make zip32 files for compatibility reasons.

Note that at the time of writing zip64 support is not ubiquitous. For example,
Libre Office doesn't seem to support files zipped with zip64, and Safari seems
to have an issue where if the first file from an zip is zip64, it _only_ unzips
that one. So while this is a niche feature, it's still a useful one in some
circumstances - to test that certain files really are zip32.