feat: async interface (via threads)
This adds a function, async_stream_unzip, that's the async version of
stream_unzip.

It's inspired by the code and discussion in the PR at
https://github.com/uktrade/stream-unzip/pull/80. However, to avoid duplication,
we have a threads-based layer to call into the existing stream_unzip function,
rather than have an almost identical copy but with async/await where needed in
it. This is similar to how https://github.com/uktrade/stream-zip provides an
async interface.