feat: add option allowed_encryption_mechanisms
Inspired by the discussion at
https://github.com/uktrade/stream-unzip/discussions/96, this adds the option
allow_encryption_mechanisms for client code to pass a list of allowed
encryption mechanisms, and so raises an exception if a file in the ZIP does not
match the allowed mechanisms.

The default is the existing behaviour - all mechanisms are allowed. However, am
tempted to only allow AES 256 by default in future releases for the default of
highest security possible.