feat: faster ZipCrypto decryption (via Rust)
This decreases the time to decrypt ZipCrypto ZIPs by approximately a factor of
10 (measured using 100MiB of random data zipped).

This uses pyo3 to make Rust code available to Python. However, recent versions
of pyo3 don't support Python 3.6 (and older don't seem to support Python 3.12
onwards), and judging the factor of 10 worth it to drop support. This is why we
dropped Python 3.6 support in an earlier commit.

Did do a few quick comparisons between
https://docs.rs/crc32fast/latest/crc32fast/ and the crc32 library used here,
https://docs.rs/crc32-v2/latest/crc32_v2/, and the one here was a bit faster.
Suspect it's because here we just call crc32 one byte at a time, and the "fast"
version has overhead per call and we never get that time back.

Co-authored-by: Uka Osim <uka.osim@digital.trade.gov.uk>
Co-authored-by: Michal Charemza <michal@charemza.name>